marine-pipeline-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.2.4
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
geoalchemy2==0.17.1
httpx==0.28.1
tenacity==9.0.0
python-multipart==0.0.20
pydantic==2.11.1
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
apscheduler==3.10.4
scikit-learn==1.6.1
joblib==1.4.2
pygbif==0.6.5
schedule==1.2.0
requests==2.32.3
python-dotenv==1.0.1

